ref(onboarding): Extract ScmPlatformFeaturesCore from ScmPlatformFeatures
Splits ScmPlatformFeatures into a reusable platform-and-features selection
core (ScmPlatformFeaturesCore) and the onboarding chrome that wraps it.
The core owns useScmPlatformDetection, useScmFeatureMeta, the manual-picker
toggle state, platform / feature derivation, the platform-selection and
feature-toggle handlers, and the step-viewed / platform-selected /
feature-toggled / change-platform-clicked analytics. The wrapper keeps the
"Create your first project" / "Choose your SDK" heading, handleContinue
logic (with the auto-create-on-skip-project-details path), and the Continue
button footer.

Shared helpers (FEATURE_DISPLAY_ORDER, getPlatformInfo, getPlatformName,
platformOptions, toSelectedSdk, shouldSuggestFramework, ResolvedPlatform)
move to a new scmPlatformHelpers.tsx so both the wrapper and the core can
import without circular dependency. The core takes an analyticsFlow prop
mirroring ScmIntegrationConnect, letting the SCM-first project creation
surface reuse just the selection mechanic and fire project_creation.scm_*
events.

The wrapper continues to call useScmPlatformDetection (React Query dedupes
with the core's call) so handleContinue can derive currentPlatformKey for
the auto-create path. The Continue button's analyticsParams drop the
"source" field since that lived in core-internal state and is already
captured by the per-selection platform_selected events.

Refs VDY-75
